+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - /proc/crypto: Ignore invalid entries (i.e. entries with missing driver or type)
 - API: do not generate error reports if parent resources are not found
 
+### Fixed
+
+- metrics: don't create error-report on closing connection nodes
+
 ## [1.30.2] - 2024-12-18
 
 ### Fixed